LayoutTests/fast/dom/microtask-detach-expected.txt

CONSOLE MESSAGE: Finished queueing tests.
CONSOLE MESSAGE: MicroTask #1 has run.
CONSOLE MESSAGE: MicroTask #0 has run.


--------
Frame: 'frame1'
--------
